refactor: 코드 리뷰 적용
코드 리뷰 링크
- https://github.com/inha-2024-capstone/chat-service/pull/3
